Makefile: desable GCC 4.9 warning about the undefined behavior in the loops
GCC 4.8 and above uses a more aggressive analysis to derive an upper
bound for the number of iterations of loops using constraints imposed
by language standards. This may cause non-conforming programs to no
longer work as expected, such as SPEC CPU 2006 464.h264ref and 416.gamess.
A new option, -fno-aggressive-loop-optimizations, was added to disable
this aggressive analysis. In some loops that have known constant number
of iterations, but undefined behavior is known to occur in the loop before
reaching or during the last iteration, GCC will warn about the undefined
behavior in the loop instead of deriving lower upper bound of the number
of iterations for the loop. The warning can be disabled
with -Wno-aggressive-loop-optimizations.